test: organize suites by behavior contracts
